Refactor follow-up questions to use new component and improve state management
